`get_obs_int()` just reads the observations and the old `get_obs_int()` is renamed `get_obs_txt()`
`get_obs_txt()` is made for people working directly with STICS input txt files. It is the one coming from SticRs (but we used it as `get_obs_int()` waiting for some improvements.
The new `get_obs_int()` can also take a plant_name argument to name the plants properly (and not just with the file name).
`get_obs_int()`  is used in `get_obs_txt()` and in `get_obs()` internally.