get_file: refactoring things already done in the lower level get_file_ function, removing some arguments. get_file_: some restructurations, for defining steps and use cases (with or whithout giving an usms.xml path), some controls added ...
in functions args, usms_filename is replaced with usms_filepath not considering for the moment searching the file in the given workspace (in the get_file function). Only the given absolute path or the relative one to the current directory are considered.

Added tests and data for get_sim, fixes for get_obs ones.
Functions help updated.